fix: Reset Core SDK when switching safes (#1259)
* fix: Only initialize the Core SDK once

* fix: Hide Send button in Address book if wrong network

* fix: Add tests for useInitSafeCoreSDK hook

* refactor: useInitSafeCoreSDK hook

* refactor: Rename initSafeSDK param

* fix: Reset core sdk when switching safes

* fix: Add safeSDK dependency for safe app modals

* Refactor: useTxSender

* fix: Adjust imports

* refactor: Use useTxSender hook in components

* fix: Handle undefined values in SignOrExecuteForm

* refactor: Extract spending limit tx to the tx sender

* fix: Failing tests

* fix: Add useTxSender functions to dependency arrays

* fix: Better type useTxSender hook

* fix: Remove ts-ignore

* fix: Remove async from undefined function
