+++ src/components/balances/AssetsTable/index.tsx
@@ -9,7 +9,7 @@ import TokenIcon from '@/components/common/TokenIcon'
 import EnhancedTable from '@/components/common/EnhancedTable'
 import TokenExplorerLink from '@/components/common/TokenExplorerLink'
 import TokenTransferModal from '@/components/tx/modals/TokenTransferModal'
-import useIsSafeOwner from '@/hooks/useIsSafeOwner'
+import useIsGranted from '@/hooks/useIsGranted'
 import Track from '@/components/common/Track'
 import { ASSETS_EVENTS } from '@/services/analytics/events/assets'
 import InfoIcon from '@/public/images/notifications/info.svg'
@@ -20,9 +20,9 @@ interface AssetsTableProps {
 
 const AssetsTable = ({ items }: AssetsTableProps): ReactElement => {
   const [selectedAsset, setSelectedAsset] = useState<string | undefined>()
-  const isSafeOwner = useIsSafeOwner()
+  const isGranted = useIsGranted()
 
-  const shouldHideActions = !isSafeOwner
+  const shouldHideActions = !isGranted
 
   const headCells = useMemo(
     () => [
