+++ src/components/settings/ContractVersion/UpdateSafeDialog.tsx
@@ -5,7 +5,7 @@ import { LATEST_SAFE_VERSION } from '@/config/constants'
 
 import TxModal from '@/components/tx/TxModal'
 
-import { createMultiSendCallOnlyTx } from '@/services/tx/txSender'
+import useTxSender from '@/hooks/useTxSender'
 import useAsync from '@/hooks/useAsync'
 
 import SignOrExecuteForm from '@/components/tx/SignOrExecuteForm'
@@ -44,13 +44,14 @@ const UpdateSafeDialog = () => {
 const ReviewUpdateSafeStep = ({ onSubmit }: { onSubmit: (txId: string) => void }) => {
   const { safe, safeLoaded } = useSafeInfo()
   const chain = useCurrentChain()
+  const { createMultiSendCallOnlyTx } = useTxSender()
 
   const [safeTx, safeTxError] = useAsync<SafeTransaction>(() => {
     if (!chain || !safeLoaded) return
 
     const txs = createUpdateSafeTxs(safe, chain)
     return createMultiSendCallOnlyTx(txs)
-  }, [chain, safe, safeLoaded])
+  }, [chain, safe, safeLoaded, createMultiSendCallOnlyTx])
 
   return (
     <SignOrExecuteForm safeTx={safeTx} onSubmit={onSubmit} error={safeTxError}>
