+++ src/components/settings/SafeModules/RemoveModule/steps/ReviewRemoveModule.tsx
@@ -1,6 +1,6 @@
 import useAsync from '@/hooks/useAsync'
 import type { SafeTransaction } from '@safe-global/safe-core-sdk-types'
-import { createRemoveModuleTx } from '@/services/tx/txSender'
+import useTxSender from '@/hooks/useTxSender'
 import SignOrExecuteForm from '@/components/tx/SignOrExecuteForm'
 import { Typography } from '@mui/material'
 import EthHashInfo from '@/components/common/EthHashInfo'
@@ -16,9 +16,10 @@ export const ReviewRemoveModule = ({
   data: RemoveModuleData
   onSubmit: (txId: string) => void
 }) => {
+  const { createRemoveModuleTx } = useTxSender()
   const [safeTx, safeTxError] = useAsync<SafeTransaction>(() => {
     return createRemoveModuleTx(data.address)
-  }, [data.address])
+  }, [data.address, createRemoveModuleTx])
 
   useEffect(() => {
     if (safeTxError) {
