+++ src/components/settings/SpendingLimits/NewSpendingLimit/steps/ReviewSpendingLimit.tsx
@@ -2,80 +2,22 @@ import { Typography, Box } from '@mui/material'
 import useBalances from '@/hooks/useBalances'
 import { useEffect, useMemo, useState } from 'react'
 import useAsync from '@/hooks/useAsync'
-import type { MetaTransactionData, SafeTransaction } from '@safe-global/safe-core-sdk-types'
+import type { SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import SignOrExecuteForm from '@/components/tx/SignOrExecuteForm'
 import EthHashInfo from '@/components/common/EthHashInfo'
 import useChainId from '@/hooks/useChainId'
 import { useSelector } from 'react-redux'
 import type { SpendingLimitState } from '@/store/spendingLimitsSlice'
 import { selectSpendingLimits } from '@/store/spendingLimitsSlice'
-import { createAddDelegateTx, createResetAllowanceTx, createSetAllowanceTx } from '@/services/tx/spendingLimitParams'
 import { getResetTimeOptions } from '@/components/transactions/TxDetails/TxData/SpendingLimits'
 import { BigNumber } from '@ethersproject/bignumber'
 import { formatVisualAmount } from '@/utils/formatters'
-import { currentMinutes, relativeTime } from '@/utils/date'
-import { getSafeSDK } from '@/hooks/coreSDK/safeCoreSDK'
-import { getSpendingLimitModuleAddress } from '@/services/contracts/spendingLimitContracts'
-import { parseUnits } from '@ethersproject/units'
-import { createMultiSendCallOnlyTx } from '@/services/tx/txSender'
+import { relativeTime } from '@/utils/date'
 import { trackEvent, SETTINGS_EVENTS } from '@/services/analytics'
 import { TokenTransferReview } from '@/components/tx/modals/TokenTransferModal/ReviewTokenTx'
 import SpendingLimitLabel from '@/components/common/SpendingLimitLabel'
-
-export const createNewSpendingLimitTx = async (
-  data: NewSpendingLimitData,
-  spendingLimits: SpendingLimitState[],
-  chainId: string,
-  tokenDecimals?: number,
-  existingSpendingLimit?: SpendingLimitState,
-) => {
-  const sdk = getSafeSDK()
-  const spendingLimitAddress = getSpendingLimitModuleAddress(chainId)
-  if (!spendingLimitAddress || !sdk) return
-
-  const txs: MetaTransactionData[] = []
-
-  const isSpendingLimitEnabled = await sdk.isModuleEnabled(spendingLimitAddress)
-  if (!isSpendingLimitEnabled) {
-    const enableModuleTx = await sdk.createEnableModuleTx(spendingLimitAddress)
-
-    const tx = {
-      to: enableModuleTx.data.to,
-      value: '0',
-      data: enableModuleTx.data.data,
-    }
-    txs.push(tx)
-  }
-
-  const existingDelegate = spendingLimits.find((spendingLimit) => spendingLimit.beneficiary === data.beneficiary)
-  if (!existingDelegate) {
-    txs.push(createAddDelegateTx(data.beneficiary, spendingLimitAddress))
-  }
-
-  if (existingSpendingLimit && existingSpendingLimit.spent !== '0') {
-    txs.push(createResetAllowanceTx(data.beneficiary, data.tokenAddress, spendingLimitAddress))
-  }
-
-  const tx = createSetAllowanceTx(
-    data.beneficiary,
-    data.tokenAddress,
-    parseUnits(data.amount, tokenDecimals).toString(),
-    parseInt(data.resetTime),
-    data.resetTime !== '0' ? currentMinutes() - 30 : 0,
-    spendingLimitAddress,
-  )
-
-  txs.push(tx)
-
-  return createMultiSendCallOnlyTx(txs)
-}
-
-export type NewSpendingLimitData = {
-  beneficiary: string
-  tokenAddress: string
-  amount: string
-  resetTime: string
-}
+import useTxSender from '@/hooks/useTxSender'
+import type { NewSpendingLimitData } from '@/services/tx/tx-sender'
 
 type Props = {
   data: NewSpendingLimitData
@@ -87,6 +29,7 @@ export const ReviewSpendingLimit = ({ data, onSubmit }: Props) => {
   const spendingLimits = useSelector(selectSpendingLimits)
   const chainId = useChainId()
   const { balances } = useBalances()
+  const { createNewSpendingLimitTx } = useTxSender()
 
   useEffect(() => {
     const existingSpendingLimit = spendingLimits.find(
@@ -107,7 +50,7 @@ export const ReviewSpendingLimit = ({ data, onSubmit }: Props) => {
 
   const [safeTx, safeTxError] = useAsync<SafeTransaction | undefined>(() => {
     return createNewSpendingLimitTx(data, spendingLimits, chainId, decimals, existingSpendingLimit)
-  }, [data, spendingLimits, chainId, decimals, existingSpendingLimit])
+  }, [data, spendingLimits, chainId, decimals, existingSpendingLimit, createNewSpendingLimitTx])
 
   const onFormSubmit = (txId: string) => {
     trackEvent({
