+++ src/components/tx/SignOrExecuteForm/SignOrExecuteForm.test.tsx
@@ -3,7 +3,8 @@ import SignOrExecuteForm from '@/components/tx/SignOrExecuteForm/index'
 import type { SafeSignature, SafeTransaction } from '@safe-global/safe-core-sdk-types'
 import * as useSafeInfoHook from '@/hooks/useSafeInfo'
 import * as useGasLimitHook from '@/hooks/useGasLimit'
-import * as txSender from '@/services/tx/txSender'
+import * as txSenderDispatch from '@/services/tx/tx-sender/dispatch'
+import * as txSender from '@/hooks/useTxSender'
 import * as wallet from '@/hooks/wallets/useWallet'
 import * as walletUtils from '@/hooks/wallets/wallets'
 import * as web3 from '@/hooks/wallets/web3'
@@ -16,6 +17,7 @@ import { Web3Provider } from '@ethersproject/providers'
 import { ethers } from 'ethers'
 import * as wrongChain from '@/hooks/useIsWrongChain'
 import * as useIsValidExecutionHook from '@/hooks/useIsValidExecution'
+import type { NullableTxSenderFunctions } from '@/hooks/useTxSender'
 
 jest.mock('@/hooks/useIsWrongChain', () => ({
   __esModule: true,
@@ -86,10 +88,10 @@ describe('SignOrExecuteForm', () => {
     jest.spyOn(web3, 'useWeb3').mockReturnValue(mockProvider)
     jest.spyOn(wrongChain, 'default').mockReturnValue(false)
     jest
-      .spyOn(txSender, 'dispatchTxProposal')
+      .spyOn(txSenderDispatch, 'dispatchTxProposal')
       .mockImplementation(jest.fn(() => Promise.resolve({ txId: '0x12' } as TransactionDetails)))
 
-    jest.spyOn(txSender, 'dispatchTxExecution').mockImplementation(jest.fn())
+    jest.spyOn(txSenderDispatch, 'dispatchTxExecution').mockImplementation(jest.fn())
     jest.spyOn(walletUtils, 'shouldUseEthSignMethod').mockImplementation(jest.fn(() => false))
   })
 
@@ -327,7 +329,7 @@ describe('SignOrExecuteForm', () => {
   })
 
   it('displays an error if execution submission fails', async () => {
-    jest.spyOn(txSender, 'dispatchTxProposal').mockImplementation(() => {
+    jest.spyOn(txSenderDispatch, 'dispatchTxProposal').mockImplementation(() => {
       throw new Error('Error while dispatching')
     })
 
@@ -391,7 +393,15 @@ describe('SignOrExecuteForm', () => {
   })
 
   it('executes a transaction', async () => {
-    const executionSpy = jest.spyOn(txSender, 'dispatchTxExecution')
+    const executionSpy = jest.fn()
+    jest.spyOn(txSender, 'default').mockImplementation(
+      () =>
+        ({
+          dispatchTxProposal: jest.fn(() => Promise.resolve({})),
+          dispatchTxExecution: executionSpy,
+        } as unknown as NullableTxSenderFunctions),
+    )
+
     const mockTx = createSafeTx()
     const result = render(<SignOrExecuteForm isExecutable={true} onSubmit={jest.fn} safeTx={mockTx} />)
 
@@ -406,8 +416,17 @@ describe('SignOrExecuteForm', () => {
 
   it('signs a transactions', async () => {
     const mockTx = createSafeTx()
-    const signSpy = jest.spyOn(txSender, 'dispatchTxSigning').mockReturnValue(Promise.resolve(mockTx))
-    const proposeSpy = jest.spyOn(txSender, 'dispatchTxProposal')
+
+    const signSpy = jest.fn(() => Promise.resolve({}))
+    const proposeSpy = jest.fn(() => Promise.resolve({}))
+    jest.spyOn(txSender, 'default').mockImplementation(
+      () =>
+        ({
+          dispatchTxSigning: signSpy,
+          dispatchTxProposal: proposeSpy,
+        } as unknown as NullableTxSenderFunctions),
+    )
+
     jest.spyOn(walletUtils, 'isSmartContractWallet').mockImplementation(() => Promise.resolve(false))
 
     const result = render(<SignOrExecuteForm onSubmit={jest.fn} safeTx={mockTx} />)
@@ -424,8 +443,17 @@ describe('SignOrExecuteForm', () => {
 
   it('smart contract wallets propose and sign new transactions on-chain', async () => {
     const mockTx = createSafeTx()
-    const onChainSignSpy = jest.spyOn(txSender, 'dispatchOnChainSigning')
-    const proposeSpy = jest.spyOn(txSender, 'dispatchTxProposal')
+
+    const onChainSignSpy = jest.fn(() => Promise.resolve({}))
+    const proposeSpy = jest.fn(() => Promise.resolve({}))
+    jest.spyOn(txSender, 'default').mockImplementation(
+      () =>
+        ({
+          dispatchOnChainSigning: onChainSignSpy,
+          dispatchTxProposal: proposeSpy,
+        } as unknown as NullableTxSenderFunctions),
+    )
+
     jest.spyOn(walletUtils, 'isSmartContractWallet').mockImplementation(() => Promise.resolve(true))
 
     const result = render(<SignOrExecuteForm onSubmit={jest.fn} safeTx={mockTx} />)
@@ -442,8 +470,17 @@ describe('SignOrExecuteForm', () => {
 
   it("smart contract wallets dont't propose, but sign existing transactions on-chain", async () => {
     const mockTx = createSafeTx()
-    const onChainSignSpy = jest.spyOn(txSender, 'dispatchOnChainSigning')
-    const proposeSpy = jest.spyOn(txSender, 'dispatchTxProposal')
+
+    const onChainSignSpy = jest.fn(() => Promise.resolve({}))
+    const proposeSpy = jest.fn(() => Promise.resolve({}))
+    jest.spyOn(txSender, 'default').mockImplementation(
+      () =>
+        ({
+          dispatchOnChainSigning: onChainSignSpy,
+          dispatchTxProposal: proposeSpy,
+        } as unknown as NullableTxSenderFunctions),
+    )
+
     jest.spyOn(walletUtils, 'isSmartContractWallet').mockImplementation(() => Promise.resolve(true))
 
     const result = render(<SignOrExecuteForm txId="0x123" onSubmit={jest.fn} safeTx={mockTx} />)
