+++ src/components/tx/modals/NftTransferModal/ReviewNftTx.tsx
@@ -6,7 +6,7 @@ import SignOrExecuteForm from '../../SignOrExecuteForm'
 import EthHashInfo from '@/components/common/EthHashInfo'
 import useAsync from '@/hooks/useAsync'
 import { createNftTransferParams } from '@/services/tx/tokenTransferParams'
-import { createTx } from '@/services/tx/txSender'
+import useTxSender from '@/hooks/useTxSender'
 import { type NftTransferParams } from '.'
 import ImageFallback from '@/components/common/ImageFallback'
 import useSafeAddress from '@/hooks/useSafeAddress'
@@ -17,13 +17,14 @@ type ReviewNftTxProps = {
 }
 
 const ReviewNftTx = ({ params, onSubmit }: ReviewNftTxProps): ReactElement => {
+  const { createTx } = useTxSender()
   const safeAddress = useSafeAddress()
   const { token } = params
 
   const [safeTx, safeTxError] = useAsync<SafeTransaction>(() => {
     const transferParams = createNftTransferParams(safeAddress, params.recipient, params.token.id, params.token.address)
     return createTx(transferParams, params.txNonce)
-  }, [safeAddress, params])
+  }, [safeAddress, params, createTx])
 
   return (
     <SignOrExecuteForm safeTx={safeTx} onSubmit={onSubmit} error={safeTxError}>
