+++ src/components/tx/modals/TokenTransferModal/ReviewMultisigTx.tsx
@@ -6,7 +6,7 @@ import SignOrExecuteForm from '@/components/tx/SignOrExecuteForm'
 import { createTokenTransferParams } from '@/services/tx/tokenTransferParams'
 import useBalances from '@/hooks/useBalances'
 import useAsync from '@/hooks/useAsync'
-import { createTx } from '@/services/tx/txSender'
+import useTxSender from '@/hooks/useTxSender'
 import EthHashInfo from '@/components/common/EthHashInfo'
 import SendFromBlock from '../../SendFromBlock'
 import type { TokenTransferModalProps } from '.'
@@ -14,6 +14,7 @@ import { TokenTransferReview } from '@/components/tx/modals/TokenTransferModal/R
 
 const ReviewMultisigTx = ({ params, onSubmit }: TokenTransferModalProps): ReactElement => {
   const { balances } = useBalances()
+  const { createTx } = useTxSender()
 
   const token = balances.items.find((item) => item.tokenInfo.address === params.tokenAddress)
   const { decimals, address } = token?.tokenInfo || {}
@@ -23,7 +24,7 @@ const ReviewMultisigTx = ({ params, onSubmit }: TokenTransferModalProps): ReactE
     if (!address || !decimals) return
     const txParams = createTokenTransferParams(params.recipient, params.amount, decimals, address)
     return createTx(txParams, params.txNonce)
-  }, [params, decimals, address])
+  }, [params, decimals, address, createTx])
 
   return (
     <SignOrExecuteForm safeTx={safeTx} onSubmit={onSubmit} error={safeTxError}>
