+++ src/components/tx/modals/TokenTransferModal/ReviewSpendingLimitTx.tsx
@@ -18,7 +18,7 @@ import useSafeInfo from '@/hooks/useSafeInfo'
 import { Errors, logError } from '@/services/exceptions'
 import ErrorMessage from '@/components/tx/ErrorMessage'
 import { useCurrentChain } from '@/hooks/useChains'
-import { dispatchSpendingLimitTxExecution } from '@/services/tx/txSender'
+import { dispatchSpendingLimitTxExecution } from '@/services/tx/tx-sender'
 import { getTxOptions } from '@/utils/transactions'
 import { MODALS_EVENTS, trackEvent } from '@/services/analytics'
