+++ src/hooks/coreSDK/useInitSafeCoreSDK.ts
@@ -1,29 +1,35 @@
-import { useEffect, useState } from 'react'
-import { useCurrentChain } from '@/hooks/useChains'
+import { useEffect } from 'react'
 import useWallet from '../wallets/useWallet'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import { initSafeSDK, setSafeSDK } from '@/hooks/coreSDK/safeCoreSDK'
+import { trackError } from '@/services/exceptions'
+import ErrorCodes from '@/services/exceptions/ErrorCodes'
+import { useAppDispatch } from '@/store'
+import { showNotification } from '@/store/notificationsSlice'
 
-export const useInitSafeCoreSDK = (): Error | null => {
-  const chain = useCurrentChain()
+export const useInitSafeCoreSDK = () => {
   const wallet = useWallet()
   const { safe, safeLoaded } = useSafeInfo()
-
-  const [error, setError] = useState<Error | null>(null)
+  const dispatch = useAppDispatch()
 
   useEffect(() => {
-    if (!safeLoaded || !chain || !wallet || chain.chainId !== wallet.chainId) {
+    if (!safeLoaded || !wallet?.provider || safe.chainId !== wallet.chainId) {
+      // If we don't reset the SDK, a previous Safe could remain in the store
+      setSafeSDK(undefined)
       return
     }
 
-    initSafeSDK(wallet.provider, wallet.chainId, safe.address.value, safe.version)
+    initSafeSDK(wallet.provider, safe.chainId, safe.address.value, safe.version)
       .then(setSafeSDK)
       .catch((e) => {
-        // If we don't reset the SDK, a previous Safe could remain in the store
-        setSafeSDK(undefined)
-        setError(e)
+        dispatch(
+          showNotification({
+            message: `The Safe SDK could not be initialized.`,
+            groupKey: 'core-sdk-init-error',
+            variant: 'error',
+          }),
+        )
+        trackError(ErrorCodes._105, (e as Error).message)
       })
-  }, [chain, wallet, safe, safeLoaded])
-
-  return error
+  }, [wallet?.provider, wallet?.chainId, safe.chainId, safe.address.value, safe.version, safeLoaded, dispatch])
 }
