+++ src/services/exceptions/ErrorCodes.ts
@@ -11,6 +11,7 @@ enum ErrorCodes {
   _101 = '101: Failed to resolve the address',
   _103 = '103: Error creating a SafeTransaction',
   _104 = '104: Invalid chain short name in the URL',
+  _105 = '105: Error initializing the Safe Core SDK',
 
   _302 = '302: Error connecting to the wallet',
   _303 = '303: Error creating pairing session',
