+++ src/utils/transactions.ts
@@ -25,7 +25,6 @@ import type { MetaTransactionData } from '@safe-global/safe-core-sdk-types/dist/
 import { OperationType } from '@safe-global/safe-core-sdk-types/dist/src/types'
 import { getGnosisSafeContractInstance } from '@/services/contracts/safeContracts'
 import extractTxInfo from '@/services/tx/extractTxInfo'
-import { createExistingTx } from '@/services/tx/txSender'
 import type { AdvancedParameters } from '@/components/tx/AdvancedParams'
 import type { TransactionOptions } from '@safe-global/safe-core-sdk-types'
 import { hasFeature } from '@/utils/chains'
@@ -139,14 +138,6 @@ export const getTxsWithDetails = (txs: Transaction[], chainId: string) => {
   return Promise.all(
     txs.map(async (tx) => {
       return await getTransactionDetails(chainId, tx.transaction.id)
-    }),
-  )
-}
-
-export const getSafeTxs = (txs: TransactionDetails[], chainId: string, safeAddress: string) => {
-  return Promise.all(
-    txs.map(async (tx) => {
-      return await createExistingTx(chainId, safeAddress, tx.txId, tx)
     }),
   )
 }
